# openapi.yaml
openapi: 3.1.0
info:
  title: Costalyx API
  version: 1.0.0
  description: >
    Multi-cloud cost intelligence platform API. This is a scaffolding
    skeleton generated from 02-DATA-MODEL.md and 03-API-CONTRACTS.md.
    Codex should extend, not replace, this structure. Additive-only changes
    once implementation begins (see 02-DATA-MODEL.md migration discipline).
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /api/v1
    description: Same-origin API v1
  - url: http://localhost:3000/api/v1
    description: Local development API v1
...
security:
  - bearerAuth: []
paths:
...
  /cost-records:
    get:
      tags: [cost-data]
      summary: List normalized cost records
      operationId: listCostRecords
      x-required-role: viewer
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Provider"
        - name: accountId
          in: query
          schema:
            type: string
            format: uuid
        - name: service
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/DateFrom"
        - $ref: "#/components/parameters/DateTo"
        - name: dimension
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated cost records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCostRecords"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /cost-records/summary:
    get:
      tags: [cost-data]
      summary: Get pre-aggregated cost totals
      operationId: getCostRecordsSummary
      x-required-role: viewer
      parameters:
        - $ref: "#/components/parameters/Provider"
        - name: accountId
          in: query
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/DateFrom"
        - $ref: "#/components/parameters/DateTo"
        - name: dimension
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Cost summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CostSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /cost-records/export:
    get:
      tags: [cost-data]
      summary: Export normalized cost records as CSV
      operationId: exportCostRecords
      x-required-role: viewer
      responses:
        "200":
          description: CSV export of normalized cost records
          content:
            text/csv:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /dimensions:
    get:
      tags: [dimensions]
      summary: List custom dimensions
      operationId: listDimensions
      x-required-role: viewer
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Paginated dimensions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedDimensions"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [dimensions]
      summary: Create an unbounded custom dimension
      operationId: createDimension
      x-required-role: analyst
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DimensionCreate"
      responses:
        "201":
          description: Dimension created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dimension"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /resource-tags:
    get:
      tags: [dimensions]
      summary: List tags for a resource
      operationId: listResourceTags
      x-required-role: viewer
      parameters:
        - name: resourceId
          in: query
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Paginated resource tags
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedResourceTags"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [dimensions]
      summary: Upsert a manual or inferred resource tag
      operationId: upsertResourceTag
      x-required-role: analyst
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceTagUpsert"
      responses:
        "201":
          description: Resource tag upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceTag"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak-issued OIDC access token.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Required on every mutating request; duplicate keys return the original result.
      schema:
        type: string
        minLength: 16
        maxLength: 128
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 25
    Provider:
      name: provider
      in: query
      schema:
        $ref: "#/components/schemas/CloudProvider"
    DateFrom:
      name: from
      in: query
      schema:
        type: string
        format: date-time
    DateTo:
      name: to
      in: query
      schema:
        type: string
        format: date-time
  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Missing or invalid bearer token
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: Authenticated caller lacks the required role
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    Conflict:
      description: Request conflicts with current state or idempotency replay
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
    Problem:
      type: object
      required: [type, title, status, detail, instance]
      properties:
        type:
          type: string
          format: uri
          examples: ["https://api.costalyx.local/errors/validation-error"]
        title:
          type: string
          examples: ["Validation Error"]
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
        requestId:
          type: string
        errors:
          type: array
          items:
            type: object
            additionalProperties: true
...
    CloudProvider:
      type: string
      enum: [aws, azure, gcp]
    LeaseType:
      type: string
      enum: [on_demand, reserved, savings_plan, spot]
...
    MoneyString:
      type: string
      pattern: '^-?[0-9]+(\.[0-9]{1,8})?$'
      examples: ["1234.5600"]
    ListMeta:
      type: object
      required: [total, page, pageSize]
      properties:
        total:
          type: integer
          minimum: 0
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 200
...
    CostRecord:
      type: object
      required:
        - id
        - provider
        - accountId
        - resourceId
        - serviceName
        - leaseType
        - hourlyRateUsd
        - usageHours
        - costTotalUsd
        - isEstimate
        - validFrom
      properties:
        id:
          type: string
          format: uuid
        provider:
          $ref: "#/components/schemas/CloudProvider"
        accountId:
          type: string
          format: uuid
        resourceId:
          type: string
        serviceName:
          type: string
        usageFamily:
          type: string
        leaseType:
          $ref: "#/components/schemas/LeaseType"
        transactionType:
          type: string
        hourlyRateUsd:
          $ref: "#/components/schemas/MoneyString"
        usageHours:
          type: string
          pattern: '^[0-9]+(\.[0-9]{1,4})?$'
        costTotalUsd:
          allOf:
            - $ref: "#/components/schemas/MoneyString"
          readOnly: true
          description: Computed at query time, never stored.
        isEstimate:
          type: boolean
        validFrom:
          type: string
          format: date-time
        validTo:
          type: ["string", "null"]
          format: date-time
        ingestedAt:
          type: string
          format: date-time
    CostSummary:
      type: object
      required: [totalCostUsd, resourceCount, untaggedCount, inactiveCount, isEstimate]
      properties:
        totalCostUsd:
          $ref: "#/components/schemas/MoneyString"
        resourceCount:
          type: integer
          minimum: 0
        untaggedCount:
          type: integer
          minimum: 0
        inactiveCount:
          type: integer
          minimum: 0
        isEstimate:
          type: boolean
...
    Dimension:
      type: object
      required: [id, orgId, name, createdBy, createdAt]
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        name:
          type: string
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    DimensionCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
...
    ResourceTag:
      type: object
      required: [resourceId, tagKey, tagValue, source]
      properties:
        resourceId:
          type: string
        tagKey:
          type: string
        tagValue:
          type: string
        source:
          type: string
          enum: [native, manual, inferred]
    ResourceTagUpsert:
      type: object
      required: [resourceId, tagKey, tagValue, source]
      properties:
        resourceId:
          type: string
        tagKey:
          type: string
        tagValue:
          type: string
        source:
          type: string
          enum: [native, manual, inferred]
...
    PaginatedCostRecords:
      allOf:
        - $ref: "#/components/schemas/PaginatedResponse"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/CostRecord"
...
    PaginatedDimensions:
      allOf:
        - $ref: "#/components/schemas/PaginatedResponse"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Dimension"
    PaginatedResourceTags:
      allOf:
        - $ref: "#/components/schemas/PaginatedResponse"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/ResourceTag"
...
    PaginatedResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: {}
        meta:
          $ref: "#/components/schemas/ListMeta"